fix(bucket): move interface to proper folder
